pyproject.toml
[tool.poetry]
name = "langgraph-chatbot"
version = "0.1.0"
description = ""
authors = ["Alonso Isidoro Román <alonsoir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.31"
langsmith = "^0.1.129"
langchain-anthropic = "^0.2.1"
anthropic = "^0.34.2"
langchain = "^0.3.1"
defusedxml = "^0.7.1"
langchain-core = "^0.3.7"
pydantic = "^2.9.2"
langchain-community = "^0.3.1"
python-dotenv = "^1.0.1"
networkx = "^3.3"
black = "^24.8.0"
tesseract = "^0.1.3"
pytesseract = "^0.3.13"
pillow = "^10.4.0"
groq = "^0.11.0"
websockets = "^13.1"
openai = "^1.51.0"
langchain-openai = "^0.2.2"
chromadb = "^0.5.11"
pdfminer-six = "^20240706"
pandas = "^2.2.3"
mutagen = "^1.47.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
